+++ frontend/src/lib/utils.ts
@@ -8,7 +8,10 @@ export function cn(...inputs: ClassValue[]) {
 
 export type WithoutChild<T> = T extends { child?: any } ? Omit<T, "child"> : T;
 export type WithoutChildrenOrChild<T> = WithoutChildren<WithoutChild<T>>;
-export type WithoutChildren<T> = T extends { children?: any } ? Omit<T, "children"> : T;
+export type WithoutChildren<T> = T extends { children?: any } ? Omit<T, "children"> : T;	
+export type WithElementRef<T, U extends HTMLElement = HTMLElement> = T & {
+  ref?: U | null;
+};
 
 const secondLen = 1000;
 const minuteLen = secondLen * 60;