+++ .github/release-notes.md
@@ -2,3 +2,4 @@
 * Added green highlight to players in your game when viewing history
 * Fixed players not being clickable in past game popup
 * Added display for known friends when clicking on player in game
+* Added automatic tag for steam friends